refactor: update remaining commands and remove recipes view
